Add a test case checking for cluster upgradeability
Currently if a cluster fails to upgrade because for example there's a
degraded operator, the failure message isn't clear. Example run:

https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_ovn-kubernetes/1583/pull-ci-openshift-ovn-kubernetes-master-e2e-aws-ovn-upgrade/1636065778761469952

This PR adds a test to check the cluster is upgradeable before beginning
the upgrade. It also adds a test case after the upgrade to check that
cluster returns to an upgradeable state (i.e. no upgrade is running
anymore, and everything comes back healthy).