fix: remove un-needed test
This test does not accuratly capture the bug it references, the matcher would match on correct reasons to mark a mc as degraded thus giving false positives. Also the test is not needed because if the condition the test is checking for were ever to arise the install would simply fail before we ever come to this state. The bug the test references also has tests so we should be fine to remove this check since it's not providing any value.

Signed-off-by: ehila <ehila@redhat.com>