virt, live-migration: add per role tests
This add final tests for VM/VMIs, restart/liveMigration and
secondary/primary UDN tests.

For primary UDN the vmi status do not present the IPs yet so the guest
is inspected, in the future the expected address can be retrieved from
the VMI status for primary UDNs too, although guest will have to be
inspected to ensure that the DHCP/NDP from OVN is working as expected.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>
Co-authored-by: Miguel Duarte Barroso <mdbarroso@redhat.com>